Move /viewcompanyresult route to the top
